nemo-service/requirements.txt
# FastAPI и веб-сервер
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
numpy==1.26.0

nemo-toolkit==2.4.1
hydra-core==1.3.2
omegaconf==2.3.0
pytorch-lightning==2.5.5
torch==2.8.0

pydub==0.25.1
